+++ sdk/communication/azure-communication-networktraversal/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0-beta.2 (2021-11-18)
 
 ### Features Added
+
+- Made User Identity an optional parameter when getting a Relay Configuration.
+- Added RouteType as optional parameter when getting a Relay Configuration so users can
+  choose the routing type protocol of the requested Relay Configuration.
 
 ### Breaking Changes
 
@@ -17,5 +21,6 @@ The first preview of the Azure Communication Relay Client has the following feat
 - get a relay configuration by creating a CommunicationRelayClient
 
 ### Features Added
+
 - Added CommunicationRelayClient in preview.
 - Added CommunicationRelayClient.getRelayConfiguration in preview.